test: Move configuration into integration folder.
