+++ test/integration/auth-allow-all.json
