+++ test/integration/auth-ldp-handler.json
